requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.31.1
semantic-kernel==1.38.0
python-dotenv==1.0.0
pydantic===2.11.0
aiofiles==23.2.1
httpx==0.27.1
mcp==1.21.1
python-json-logger==2.0.7